requirements.txt
pip
django
djangorestframework
mysqlclient
git
